automation/requirements.txt
requests
supabase
beautifulsoup4
python-dotenv
google-generativeai
